fix: UiDevice Extension 메서드 위치 변경
